zlp 取消订单将支付的红包金额还回

+++ workspace/bbyiya-basic/src/main/java/com/bbyiya/service/impl/BasePayServiceImpl.java
@@ -169,7 +169,7 @@ public class BasePayServiceImpl implements IBasePayService{
 									log.setOrderid(payId);
 									accountslogsMapper.insert(log);
 									//更新钱包的冻结金额
-									accounts.setFreezecashamount(accounts.getFreezecashamount().doubleValue()-payOrder.getWalletamount().intValue());
+									accounts.setFreezecashamount(accounts.getFreezecashamount().doubleValue()-payOrder.getWalletamount().doubleValue());
 									accountsMapper.updateByPrimaryKeySelective(accounts);
 								}
 								//更新订单的状态---
